notes fixing
