+++ .gitignore
@@ -1,5 +1,7 @@
 # Directories
 node_modules/
+build/
+logs/
 
 #IDE generated
 .idea/
