+++ src/version/plugins/Feeds/RssItemsList.js
@@ -9,7 +9,8 @@ import Typography from "@material-ui/core/Typography";
  * This component returns synopsis list
  *
  * @author Bogdan Shcherban <bsc@piogroup.net>
- * @param {array}  items
+ * @param {shape} classes
+ * @param {array} items
  * @constructor
  */
 const ItemsList = ({ classes, items }) => {
@@ -22,8 +23,7 @@ const ItemsList = ({ classes, items }) => {
     } else {
         return (
             <List className={classes.list}>
-                {items.slice(0, 4).map((item, key) => {
-                    return (
+                {items.slice(0, 4).map((item, key) => (
                         <a href={item.link} target="_blank">
                             <ListItem key={key} button divider>
                                 <Typography noWrap={true} className={classes.feedsItem}>
@@ -31,8 +31,8 @@ const ItemsList = ({ classes, items }) => {
                                 </Typography>
                             </ListItem>
                         </a>
-                    );
-                })}
+                    )
+                )}
             </List>
         );
     }
