+++ src/version/sagas/feedsRssSagas.js
@@ -3,12 +3,10 @@ import { takeEvery, put } from 'redux-saga/effects';
 import { FEEDS_RSS_ACTION, feedsRssAction } from "../actions/feedsRssAction";
 import { getRssFeedsListFromXML } from "../plugins/Feeds/rss-helper";
 
-const CORS_PROXY = "https://cors-anywhere.herokuapp.com/";
-
 export default takeEvery(FEEDS_RSS_ACTION.REQUEST, function*(action) {
     const feedsUrl = action.rssFeedUrl;
     const sourceId = action.sourceId;
-    const url = CORS_PROXY + feedsUrl.replace("http://", "").replace("https://", "");
+    const url = "https://cors-anywhere.herokuapp.com/" + feedsUrl.replace("http://", "").replace("https://", "");
     try {
         const result = yield fetch(url)
             .then(res => res.text())
